+++ pyjinhx/builtins/ui/pjx_panel/__init__.py
@@ -1,4 +1,0 @@
-from .pjx_panel import PJXPanel
-from .pjx_panel_trigger import PJXPanelTrigger
-
-__all__ = ["PJXPanel", "PJXPanelTrigger"]